@@ -2,29 +2,29 @@ import { createSlice } from '@reduxjs/toolkit';
 import { loginService } from '../services/login';
 import { blogsService } from '../services/blogs';
 
-const userSlice = createSlice({
-  name: 'user',
+const currentUserSlice = createSlice({
+  name: 'currentUser',
   initialState: null,
   reducers: {
-    setUser(state, action) {
+    setCurrentUser(state, action) {
       return action.payload;
     },
   },
 });
 
-export const { setUser } = userSlice.actions;
+export const { setCurrentUser } = currentUserSlice.actions;
 
-export const initializeUser = (userToLogin) => {
+export const initializeCurrentUser = (userToLogin) => {
   return async (dispatch) => {
     try {
       const user = await loginService.loginBlogs(userToLogin);
       window.localStorage.setItem('loggedBlogsAppUser', JSON.stringify(user));
       blogsService.setToken(user.token);
-      dispatch(setUser(user));
+      dispatch(setCurrentUser(user));
     } catch (e) {
       console.error(e);
     }
   };
 };
 
-export default userSlice.reducer;
+export default currentUserSlice.reducer;
